Refactor the retry policies
